requirements.txt
aiofiles==0.5.0
certifi==2020.4.5.1
chardet==3.0.4
h11==0.8.1
h2==3.2.0
hpack==3.0.0
hstspreload==2020.5.5
httptools==0.1.1
httpx==0.9.3
hyperframe==5.2.0
idna==2.9
multidict==4.7.5
pyserial==3.4
rfc3986==1.4.0
sanic==19.12.2
sniffio==1.1.0
ujson==2.0.3
uvloop==0.14.0
websockets==8.1
